fix: move CommitLines dir and update export map
