+++ app/src/lib/commit/CommitList.svelte
@@ -19,8 +19,7 @@
 		getRemoteCommits
 	} from '$lib/vbranches/contexts';
 	import { BaseBranch, Branch } from '$lib/vbranches/types';
-	import LineGroup from '@gitbutler/ui/CommitLines/LineGroup.svelte';
-	import { LineManagerFactory } from '@gitbutler/ui/CommitLines/lineManager';
+	import { LineManagerFactory, LineGroup } from '@gitbutler/ui';
 	import { goto } from '$app/navigation';
 
 	export let isUnapplied: boolean;
