+++ packages/ui/src/lib/index.ts
@@ -1,0 +1,21 @@
+// This file defines the short path imports for the package (ex: @gitbutler/ui/*)
+
+// export type {
+// 	Color,
+// 	Style,
+// 	CommitNode,
+// 	BaseNode,
+// 	Line,
+// 	LineGroup as LineGroupSettings,
+// 	Author,
+// 	CommitData
+// } from './components/CommitLines/types.ts';
+
+export { LineManager, LineManagerFactory } from './components/CommitLines/lineManager';
+export { default as LineGroup } from './components/CommitLines/LineGroup.svelte';
+export { default as Line } from './components/CommitLines/Line.svelte';
+export { default as CommitNode } from './components/CommitLines/CommitNode.svelte';
+export { default as Cell } from './components/CommitLines/Cell.svelte';
+export { default as BaseNode } from './components/CommitLines/BaseNode.svelte';
+export { default as Fork } from './components/CommitLines/Cell/Fork.svelte';
+export { default as Straight } from './components/CommitLines/Cell/Straight.svelte';
