+++ packages/ui/src/lib/types.ts
@@ -1,0 +1,10 @@
+export type {
+	Color,
+	Style,
+	CommitNode,
+	BaseNode,
+	Line,
+	LineGroup as LineGroupSettings,
+	Author,
+	CommitData
+} from './components/CommitLines/types.ts';
